UX-408 Add readme, move helper functions, rename props
