+++ packages/DatePicker/README.md
@@ -1,0 +1,48 @@
+## Date picker
+
+### Installation
+
+`> npm install --save @paprika/date-picker`
+or
+`> yarn add @paprika/date-picker`
+
+### Usage
+
+Please use `<L10n />` component to wrap `<DatePicker />` or you application.
+
+For a basic date picker
+
+```js
+import DatePicker from "@paprika/date-picker";
+
+<DatePicker onChange={changeHandler} />;
+```
+
+Customized input
+
+```js
+import DatePicker from "@paprika/date-picker";
+
+<DatePicker onChange={changeHandler}>
+  <DatePicker.Input placeholder="Date..." />
+</DatePicker>;
+```
+
+### Props
+
+- `children`
+- `date` Moment object
+- `dataFormat`
+- `isDisabled`
+- `isReadOnly`
+- `onChange` (Required)
+
+#### DatePicker.Calendar
+
+#### DatePicker.Input
+
+- `a11yText`
+- `placeholder`
+- `size` `["small", "medium", "large"]`
+
+[More detail about these props](https://github.com/acl-services/paprika/blob/master/packages/DatePicker/src/DatePicker.js)
