+++ packages/DatePicker/src/helpers.js
@@ -11,19 +11,3 @@ export function extractChildrenProps(children) {
 
   return overrides;
 }
-
-export function useDebounce(value, delay) {
-  const [debouncedValue, setDebouncedValue] = React.useState(value);
-
-  React.useEffect(() => {
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-
-    return () => {
-      clearTimeout(handler);
-    };
-  }, [value, delay]);
-
-  return debouncedValue;
-}
