+++ packages/helpers/src/dom/isElementContainsFocus.js
@@ -1,0 +1,4 @@
+export default function isElementContainsFocus(element) {
+  if (document.activeElement === element) return true;
+  return element.contains(document.activeElement);
+}
